refactor: Merge Crad Components
