--- src/ui/EditableCard/index.js
@@ -1,1 +1,0 @@
-export * from "./EditableCard";
