--- src/ui/ReadOnlyCard/index.js
@@ -1,1 +1,0 @@
-export * from "./ReadOnlyCard";
